addded support for pre-post CD step execution
