+++ App.go
@@ -19,7 +19,6 @@ package main
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
 	_ "github.com/aws/aws-sdk-go/aws"
 	"github.com/devtron-labs/ci-runner/helper"
@@ -28,13 +27,9 @@ import (
 	"log"
 	"os"
 	"os/signal"
-	"path/filepath"
 	"runtime/debug"
-	"strconv"
-	"strings"
 	"sync"
 	"syscall"
-	"time"
 )
 
 var handleOnce sync.Once
@@ -90,459 +85,18 @@ func processEvent(args string) {
 		handleCleanup(*ciCdRequest, &abortErrorCode, util.Source_Signal)
 	}()
 
-	defer handleCleanup(*ciCdRequest, &exitCode, util.Source_Defer)
 	logLevel := os.Getenv("LOG_LEVEL")
 	if logLevel == "" || logLevel == "DEBUG" {
 		log.Println(util.DEVTRON, " ci-cd request details -----> ", args)
 	}
 
+	defer handleCleanup(*ciCdRequest, &exitCode, util.Source_Defer)
 	if ciCdRequest.Type == util.CIEVENT {
-
-		ciRequest := ciCdRequest.CiRequest
-		artifactUploaded, err := runCIStages(ciCdRequest)
-		log.Println(util.DEVTRON, artifactUploaded, err)
-		var artifactUploadErr error
-		if !artifactUploaded {
-			artifactUploaded, artifactUploadErr = helper.ZipAndUpload(ciRequest.BlobStorageConfigured, ciCdRequest.CiRequest.BlobStorageS3Config, ciCdRequest.CiRequest.CiArtifactFileName, ciCdRequest.CiRequest.CloudProvider, ciCdRequest.CiRequest.AzureBlobConfig, ciCdRequest.CiRequest.GcpBlobConfig)
-		}
-
-		if err != nil {
-			var stageError *helper.CiStageError
-			log.Println(util.DEVTRON, err)
-			if errors.As(err, &stageError) {
-				exitCode = util.CiStageFailErrorCode
-				return
-			}
-			exitCode = util.DefaultErrorCode
-			return
-		}
-
-		if artifactUploadErr != nil {
-			log.Println(util.DEVTRON, artifactUploadErr)
-			exitCode = util.DefaultErrorCode
-			return
-		}
-
-		// sync cache
-		log.Println(util.DEVTRON, " cache-push")
-		err = helper.SyncCache(ciRequest)
-		if err != nil {
-			log.Println(err)
-			exitCode = util.DefaultErrorCode
-			return
-		}
-		log.Println(util.DEVTRON, " /cache-push")
+		HandleCIEvent(ciCdRequest, &exitCode)
 	} else {
-		err = runCDStages(ciCdRequest)
-		artifactUploadErr := collectAndUploadCDArtifacts(ciCdRequest.CdRequest)
-		if err != nil || artifactUploadErr != nil {
-			log.Println(err)
-			exitCode = util.DefaultErrorCode
-			return
-		}
+		HandleCDEvent(err, ciCdRequest, &exitCode)
 	}
 	return
-}
-
-func collectAndUploadCDArtifacts(cdRequest *helper.CdRequest) error {
-	artifactFiles := make(map[string]string)
-	var allTasks []*helper.Task
-	if cdRequest.TaskYaml != nil {
-		for _, pc := range cdRequest.TaskYaml.CdPipelineConfig {
-			for _, t := range append(pc.BeforeTasks, pc.AfterTasks...) {
-				allTasks = append(allTasks, t)
-			}
-		}
-	}
-	for _, task := range allTasks {
-		if task.RunStatus {
-			if _, err := os.Stat(task.OutputLocation); os.IsNotExist(err) { // Ignore if no file/folder
-				log.Println(util.DEVTRON, "artifact not found ", err)
-				continue
-			}
-			artifactFiles[task.Name] = task.OutputLocation
-		}
-	}
-	log.Println(util.DEVTRON, " artifacts", artifactFiles)
-	return helper.UploadArtifact(cdRequest.BlobStorageConfigured, artifactFiles, cdRequest.BlobStorageS3Config, cdRequest.ArtifactFileName, cdRequest.CloudProvider, cdRequest.AzureBlobConfig, cdRequest.GcpBlobConfig)
-}
-
-func getGlobalEnvVariables(cicdRequest *helper.CiCdTriggerEvent) (map[string]string, error) {
-	envs := make(map[string]string)
-	envs["WORKING_DIRECTORY"] = util.WORKINGDIR
-	if cicdRequest.Type == util.CIEVENT {
-		image, err := helper.BuildDockerImagePath(cicdRequest.CiRequest)
-		if err != nil {
-			return nil, err
-		}
-		envs["DOCKER_IMAGE_TAG"] = cicdRequest.CiRequest.DockerImageTag
-		envs["DOCKER_REPOSITORY"] = cicdRequest.CiRequest.DockerRepository
-		envs["DOCKER_REGISTRY_URL"] = cicdRequest.CiRequest.DockerRegistryURL
-		envs["APP_NAME"] = cicdRequest.CiRequest.AppName
-		envs["TRIGGER_BY_AUTHOR"] = cicdRequest.CiRequest.TriggerByAuthor
-		envs["DOCKER_IMAGE"] = image
-
-		//adding GIT_MATERIAL_REQUEST in env for semgrep plugin
-		CiMaterialRequestArr := ""
-		if cicdRequest.CiRequest.CiProjectDetails != nil {
-			for _, ciProjectDetail := range cicdRequest.CiRequest.CiProjectDetails {
-				GitRepoSplit := strings.Split(ciProjectDetail.GitRepository, "/")
-				GitRepoName := ""
-				if len(GitRepoSplit) > 0 {
-					GitRepoName = strings.Split(GitRepoSplit[len(GitRepoSplit)-1], ".")[0]
-				}
-				CiMaterialRequestArr = CiMaterialRequestArr +
-					fmt.Sprintf("%s,%s,%s,%s|", GitRepoName, ciProjectDetail.CheckoutPath, ciProjectDetail.SourceValue, ciProjectDetail.CommitHash)
-			}
-		}
-		envs["GIT_MATERIAL_REQUEST"] = CiMaterialRequestArr // GIT_MATERIAL_REQUEST will be of form "<repoName>/<checkoutPath>/<BranchName>/<CommitHash>"
-		fmt.Println(envs["GIT_MATERIAL_REQUEST"])
-
-		// setting extraEnvironmentVariables
-		for k, v := range cicdRequest.CiRequest.ExtraEnvironmentVariables {
-			envs[k] = v
-		}
-	} else {
-		envs["DOCKER_IMAGE"] = cicdRequest.CdRequest.CiArtifactDTO.Image
-		envs["DEPLOYMENT_RELEASE_ID"] = strconv.Itoa(cicdRequest.CdRequest.DeploymentReleaseCounter)
-		envs["DEPLOYMENT_UNIQUE_ID"] = strconv.Itoa(cicdRequest.CdRequest.WorkflowRunnerId)
-		envs["DEVTRON_CD_TRIGGERED_BY"] = cicdRequest.CdRequest.DeploymentTriggeredBy
-		envs["DEVTRON_CD_TRIGGER_TIME"] = cicdRequest.CdRequest.DeploymentTriggerTime.String()
-		for k, v := range cicdRequest.CdRequest.ExtraEnvironmentVariables {
-			envs[k] = v
-		}
-	}
-	return envs, nil
-}
-
-func getSystemEnvVariables() map[string]string {
-	envs := make(map[string]string)
-	//get all environment variables
-	envVars := os.Environ()
-	for _, envVar := range envVars {
-		a := strings.Split(envVar, "=")
-		envs[a[0]] = a[1]
-	}
-	return envs
-}
-
-func sendFailureNotification(failureMessage string, ciRequest *helper.CiRequest,
-	digest string, image string, ciMetrics helper.CIMetrics,
-	artifactUploaded bool, err error) (bool, error) {
-	e := helper.SendEvents(ciRequest, digest, image, ciMetrics, artifactUploaded, failureMessage)
-	if e != nil {
-		log.Println(e)
-		return artifactUploaded, e
-	}
-	return artifactUploaded, &helper.CiStageError{Err: err}
-}
-
-type CiFailReason string
-
-const (
-	PreCi  CiFailReason = "Pre-CI task failed: "
-	PostCi CiFailReason = "Post-CI task failed: "
-	Build  CiFailReason = "Docker build failed"
-	Push   CiFailReason = "Docker push failed"
-	Scan   CiFailReason = "Image scan failed"
-)
-
-func runCIStages(ciCdRequest *helper.CiCdTriggerEvent) (artifactUploaded bool, err error) {
-
-	var metrics helper.CIMetrics
-	start := time.Now()
-	metrics.TotalStartTime = start
-	artifactUploaded = false
-	err = os.Chdir("/")
-	if err != nil {
-		return artifactUploaded, err
-	}
-
-	if _, err := os.Stat(util.WORKINGDIR); os.IsNotExist(err) {
-		_ = os.Mkdir(util.WORKINGDIR, os.ModeDir)
-	}
-
-	// Get ci cache
-	log.Println(util.DEVTRON, " cache-pull")
-	start = time.Now()
-	metrics.CacheDownStartTime = start
-	err = helper.GetCache(ciCdRequest.CiRequest)
-	metrics.CacheDownDuration = time.Since(start).Seconds()
-	if err != nil {
-		return artifactUploaded, err
-	}
-	log.Println(util.DEVTRON, " /cache-pull")
-
-	err = os.Chdir(util.WORKINGDIR)
-	if err != nil {
-		return artifactUploaded, err
-	}
-	// git handling
-	log.Println(util.DEVTRON, " git")
-	err = helper.CloneAndCheckout(ciCdRequest.CiRequest.CiProjectDetails)
-	if err != nil {
-		log.Println(util.DEVTRON, "clone err: ", err)
-		return artifactUploaded, err
-	}
-	log.Println(util.DEVTRON, " /git")
-
-	// Start docker daemon
-	log.Println(util.DEVTRON, " docker-build")
-
-	helper.StartDockerDaemon(ciCdRequest.CiRequest.DockerConnection, ciCdRequest.CiRequest.DockerRegistryURL, ciCdRequest.CiRequest.DockerCert, ciCdRequest.CiRequest.DefaultAddressPoolBaseCidr, ciCdRequest.CiRequest.DefaultAddressPoolSize, ciCdRequest.CiRequest.CiBuildDockerMtuValue)
-	scriptEnvs, err := getGlobalEnvVariables(ciCdRequest)
-	if err != nil {
-		return artifactUploaded, err
-	}
-	// Get devtron-ci yaml
-	yamlLocation := ciCdRequest.CiRequest.CheckoutPath
-	log.Println(util.DEVTRON, "devtron-ci yaml location ", yamlLocation)
-	taskYaml, err := helper.GetTaskYaml(yamlLocation)
-	if err != nil {
-		return artifactUploaded, err
-	}
-	ciCdRequest.CiRequest.TaskYaml = taskYaml
-	ciBuildConfigBean := ciCdRequest.CiRequest.CiBuildConfig
-	if ciBuildConfigBean != nil && ciBuildConfigBean.CiBuildType == helper.MANAGED_DOCKERFILE_BUILD_TYPE {
-		err = makeDockerfile(ciBuildConfigBean.DockerBuildConfig, ciCdRequest.CiRequest.CheckoutPath)
-		if err != nil {
-			return artifactUploaded, err
-		}
-	}
-
-	refStageMap := make(map[int][]*helper.StepObject)
-	for _, ref := range ciCdRequest.CiRequest.RefPlugins {
-		refStageMap[ref.Id] = ref.Steps
-	}
-
-	var preeCiStageOutVariable map[int]map[string]*helper.VariableObject
-	var step *helper.StepObject
-	var preCiDuration float64
-	start = time.Now()
-	metrics.PreCiStartTime = start
-	buildSkipEnabled := ciBuildConfigBean != nil && ciBuildConfigBean.CiBuildType == helper.BUILD_SKIP_BUILD_TYPE
-	if len(ciCdRequest.CiRequest.PreCiSteps) > 0 {
-		if !buildSkipEnabled {
-			util.LogStage("running PRE-CI steps")
-		}
-		// run pre artifact processing
-		preeCiStageOutVariable, step, err = RunCiSteps(STEP_TYPE_PRE, ciCdRequest.CiRequest.PreCiSteps, refStageMap, scriptEnvs, nil)
-		preCiDuration = time.Since(start).Seconds()
-		if err != nil {
-			log.Println(err)
-			return sendFailureNotification(string(PreCi)+step.Name, ciCdRequest.CiRequest, "", "", metrics, artifactUploaded, err)
-
-		}
-	}
-	metrics.PreCiDuration = preCiDuration
-	var dest string
-	if !buildSkipEnabled {
-		util.LogStage("Build")
-		// build
-		start = time.Now()
-		metrics.BuildStartTime = start
-		dest, err = helper.BuildArtifact(ciCdRequest.CiRequest) //TODO make it skipable
-		metrics.BuildDuration = time.Since(start).Seconds()
-		if err != nil {
-			// code-block starts : run post-ci which are enabled to run on ci fail
-			postCiStepsToTriggerOnCiFail := getPostCiStepToRunOnCiFail(ciCdRequest.CiRequest.PostCiSteps)
-			if len(postCiStepsToTriggerOnCiFail) > 0 {
-				util.LogStage("Running POST-CI steps which are enabled to RUN even on CI FAIL")
-				// build success will always be false
-				scriptEnvs[util.ENV_VARIABLE_BUILD_SUCCESS] = "false"
-				// run post artifact processing
-				RunCiSteps(STEP_TYPE_POST, postCiStepsToTriggerOnCiFail, refStageMap, scriptEnvs, preeCiStageOutVariable)
-			}
-			// code-block ends
-			return sendFailureNotification(string(Build), ciCdRequest.CiRequest, "", "", metrics, artifactUploaded, err)
-		}
-		log.Println(util.DEVTRON, " /Build")
-	}
-	var postCiDuration float64
-	start = time.Now()
-	metrics.PostCiStartTime = start
-	if len(ciCdRequest.CiRequest.PostCiSteps) > 0 {
-		util.LogStage("running POST-CI steps")
-		// sending build success as true always as post-ci triggers only if ci gets success
-		scriptEnvs[util.ENV_VARIABLE_BUILD_SUCCESS] = "true"
-		// run post artifact processing
-		_, step, err = RunCiSteps(STEP_TYPE_POST, ciCdRequest.CiRequest.PostCiSteps, refStageMap, scriptEnvs, preeCiStageOutVariable)
-		postCiDuration = time.Since(start).Seconds()
-		if err != nil {
-			return sendFailureNotification(string(PostCi)+step.Name, ciCdRequest.CiRequest, "", "", metrics, artifactUploaded, err)
-		}
-	}
-	metrics.PostCiDuration = postCiDuration
-	var digest string
-
-	if !buildSkipEnabled {
-		isBuildX := ciBuildConfigBean != nil && ciBuildConfigBean.DockerBuildConfig != nil && ciBuildConfigBean.DockerBuildConfig.TargetPlatform != ""
-		if isBuildX {
-			digest, err = helper.ExtractDigestForBuildx(dest)
-		} else {
-			util.LogStage("docker push")
-			// push to dest
-			log.Println(util.DEVTRON, " docker-push")
-			err = helper.PushArtifact(dest)
-			if err != nil {
-				return sendFailureNotification(string(Push), ciCdRequest.CiRequest, digest, dest, metrics, artifactUploaded, err)
-			}
-			digest, err = helper.ExtractDigestUsingPull(dest)
-		}
-	}
-
-	if err != nil {
-		return artifactUploaded, err
-	}
-	log.Println(util.DEVTRON, " /docker-push")
-
-	log.Println(util.DEVTRON, " artifact-upload")
-
-	artifactUploaded, err = helper.ZipAndUpload(ciCdRequest.CiRequest.BlobStorageConfigured, ciCdRequest.CiRequest.BlobStorageS3Config, ciCdRequest.CiRequest.CiArtifactFileName, ciCdRequest.CiRequest.CloudProvider, ciCdRequest.CiRequest.AzureBlobConfig, ciCdRequest.CiRequest.GcpBlobConfig)
-
-	if err != nil {
-		return artifactUploaded, err
-	}
-	//else {
-	//	artifactUploaded = true
-	//}
-	log.Println(util.DEVTRON, " /artifact-upload")
-
-	// scan only if ci scan enabled
-	if ciCdRequest.CiRequest.ScanEnabled {
-		util.LogStage("IMAGE SCAN")
-		log.Println(util.DEVTRON, " /image-scanner")
-		scanEvent := &helper.ScanEvent{Image: dest, ImageDigest: digest, PipelineId: ciCdRequest.CiRequest.PipelineId, UserId: ciCdRequest.CiRequest.TriggeredBy}
-		scanEvent.DockerRegistryId = ciCdRequest.CiRequest.DockerRegistryId
-		err = helper.SendEventToClairUtility(scanEvent)
-		if err != nil {
-			log.Println(err)
-			return sendFailureNotification(string(Scan), ciCdRequest.CiRequest, digest, dest, metrics, artifactUploaded, err)
-
-		}
-		log.Println(util.DEVTRON, " /image-scanner")
-	}
-
-	log.Println(util.DEVTRON, " event")
-	metrics.TotalDuration = time.Since(metrics.TotalStartTime).Seconds()
-
-	err = helper.SendEvents(ciCdRequest.CiRequest, digest, dest, metrics, artifactUploaded, "")
-	if err != nil {
-		log.Println(err)
-		return artifactUploaded, err
-	}
-	log.Println(util.DEVTRON, " /event")
-
-	err = helper.StopDocker()
-	if err != nil {
-		log.Println("err", err)
-		return artifactUploaded, err
-	}
-	return artifactUploaded, nil
-}
-
-func getPostCiStepToRunOnCiFail(postCiSteps []*helper.StepObject) []*helper.StepObject {
-	var postCiStepsToTriggerOnCiFail []*helper.StepObject
-	if len(postCiSteps) > 0 {
-		for _, postCiStep := range postCiSteps {
-			if postCiStep.TriggerIfParentStageFail {
-				postCiStepsToTriggerOnCiFail = append(postCiStepsToTriggerOnCiFail, postCiStep)
-			}
-		}
-	}
-	return postCiStepsToTriggerOnCiFail
-}
-
-func makeDockerfile(config *helper.DockerBuildConfig, checkoutPath string) error {
-	dockerfileContent := config.DockerfileContent
-	dockerfilePath := filepath.Join(util.WORKINGDIR, checkoutPath, "./Dockerfile")
-	f, err := os.Create(dockerfilePath)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-	_, err = f.WriteString(dockerfileContent)
-	return err
-}
-
-func runCDStages(cicdRequest *helper.CiCdTriggerEvent) error {
-	err := os.Chdir("/")
-	if err != nil {
-		return err
-	}
-
-	if _, err := os.Stat(util.WORKINGDIR); os.IsNotExist(err) {
-		_ = os.Mkdir(util.WORKINGDIR, os.ModeDir)
-	}
-	err = os.Chdir(util.WORKINGDIR)
-	if err != nil {
-		return err
-	}
-	// git handling
-	log.Println(util.DEVTRON, " git")
-	err = helper.CloneAndCheckout(cicdRequest.CdRequest.CiProjectDetails)
-	if err != nil {
-		log.Println(util.DEVTRON, "clone err: ", err)
-		return err
-	}
-	log.Println(util.DEVTRON, " /git")
-
-	// Start docker daemon
-	log.Println(util.DEVTRON, " docker-start")
-	helper.StartDockerDaemon(cicdRequest.CdRequest.DockerConnection, cicdRequest.CdRequest.DockerRegistryURL, cicdRequest.CdRequest.DockerCert, cicdRequest.CdRequest.DefaultAddressPoolBaseCidr, cicdRequest.CdRequest.DefaultAddressPoolSize, -1)
-
-	err = helper.DockerLogin(&helper.DockerCredentials{
-		DockerUsername:     cicdRequest.CdRequest.DockerUsername,
-		DockerPassword:     cicdRequest.CdRequest.DockerPassword,
-		AwsRegion:          cicdRequest.CdRequest.AwsRegion,
-		AccessKey:          cicdRequest.CdRequest.AccessKey,
-		SecretKey:          cicdRequest.CdRequest.SecretKey,
-		DockerRegistryURL:  cicdRequest.CdRequest.DockerRegistryURL,
-		DockerRegistryType: cicdRequest.CdRequest.DockerRegistryType,
-	})
-	if err != nil {
-		return err
-	}
-	// Get devtron-cd yaml
-	taskYaml, err := helper.ToTaskYaml([]byte(cicdRequest.CdRequest.StageYaml))
-	if err != nil {
-		log.Println(err)
-		return err
-	}
-	cicdRequest.CdRequest.TaskYaml = taskYaml
-
-	// run post artifact processing
-	log.Println(util.DEVTRON, " stage yaml", taskYaml)
-	var tasks []*helper.Task
-	for _, t := range taskYaml.CdPipelineConfig {
-		tasks = append(tasks, t.BeforeTasks...)
-		tasks = append(tasks, t.AfterTasks...)
-	}
-
-	scriptEnvs, err := getGlobalEnvVariables(cicdRequest)
-	if err != nil {
-		return err
-	}
-	err = RunCdStageTasks(tasks, scriptEnvs)
-	if err != nil {
-		return err
-	}
-	// dry run flag indicates that ci runner image is being run from external helm chart
-	if !cicdRequest.CdRequest.IsDryRun {
-		log.Println(util.DEVTRON, " event")
-		err = helper.SendCDEvent(cicdRequest.CdRequest)
-		if err != nil {
-			log.Println(err)
-			return err
-		}
-		log.Println(util.DEVTRON, " /event")
-	}
-	err = helper.StopDocker()
-	if err != nil {
-		log.Println("err", err)
-		return err
-	}
-	return nil
 }
 
 func uploadLogs(event helper.CiCdTriggerEvent, exitCode *int) {
