+++ StageExecutor.go
@@ -36,32 +36,32 @@ const (
 	STEP_TYPE_REF_PLUGIN StepType = "REF_PLUGIN"
 )
 
-func RunCiSteps(stepType StepType, steps []*helper.StepObject, refStageMap map[int][]*helper.StepObject, globalEnvironmentVariables map[string]string, preeCiStageVariable map[int]map[string]*helper.VariableObject) (outVars map[int]map[string]*helper.VariableObject, failedStep *helper.StepObject, err error) {
+func RunCiCdSteps(stepType StepType, steps []*helper.StepObject, refStageMap map[int][]*helper.StepObject, globalEnvironmentVariables map[string]string, preeCiStageVariable map[int]map[string]*helper.VariableObject) (outVars map[int]map[string]*helper.VariableObject, failedStep *helper.StepObject, err error) {
 	/*if stageType == STEP_TYPE_POST {
 		postCiStageVariable = make(map[int]map[string]*VariableObject) // [stepId]name[]value
 	}*/
 	stageVariable := make(map[int]map[string]*helper.VariableObject)
-	for i, ciStep := range steps {
+	for i, step := range steps {
 		var vars []*helper.VariableObject
 		if stepType == STEP_TYPE_REF_PLUGIN {
-			vars, err = deduceVariables(ciStep.InputVars, globalEnvironmentVariables, nil, nil, stageVariable)
+			vars, err = deduceVariables(step.InputVars, globalEnvironmentVariables, nil, nil, stageVariable)
 		} else {
-			log.Printf("running step : %s\n", ciStep.Name)
+			log.Printf("running step : %s\n", step.Name)
 			if stepType == STEP_TYPE_PRE {
-				vars, err = deduceVariables(ciStep.InputVars, globalEnvironmentVariables, stageVariable, nil, nil)
+				vars, err = deduceVariables(step.InputVars, globalEnvironmentVariables, stageVariable, nil, nil)
 			} else if stepType == STEP_TYPE_POST {
-				vars, err = deduceVariables(ciStep.InputVars, globalEnvironmentVariables, preeCiStageVariable, stageVariable, nil)
+				vars, err = deduceVariables(step.InputVars, globalEnvironmentVariables, preeCiStageVariable, stageVariable, nil)
 			}
 		}
 		if err != nil {
-			return nil, ciStep, err
+			return nil, step, err
 		}
-		ciStep.InputVars = vars
+		step.InputVars = vars
 
 		//variables with empty value
 		var emptyVariableList []string
 		scriptEnvs := make(map[string]string)
-		for _, v := range ciStep.InputVars {
+		for _, v := range step.InputVars {
 			scriptEnvs[v.Name] = v.Value
 			if len(v.Value) == 0 {
 				emptyVariableList = append(emptyVariableList, v.Name)
@@ -70,81 +70,81 @@ func RunCiSteps(stepType StepType, steps []*helper.StepObject, refStageMap map[i
 		if stepType == STEP_TYPE_PRE || stepType == STEP_TYPE_POST {
 			log.Println(fmt.Sprintf("variables with empty value : %v", emptyVariableList))
 		}
-		if len(ciStep.TriggerSkipConditions) > 0 {
-			shouldTrigger, err := helper.ShouldTriggerStage(ciStep.TriggerSkipConditions, ciStep.InputVars)
+		if len(step.TriggerSkipConditions) > 0 {
+			shouldTrigger, err := helper.ShouldTriggerStage(step.TriggerSkipConditions, step.InputVars)
 			if err != nil {
 				log.Println(err)
-				return nil, ciStep, err
+				return nil, step, err
 			}
 			if !shouldTrigger {
-				log.Printf("skipping %s as per pass Condition\n", ciStep.Name)
+				log.Printf("skipping %s as per pass Condition\n", step.Name)
 				continue
 			}
 		}
 
 		var outVars []string
-		for _, outVar := range ciStep.OutputVars {
+		for _, outVar := range step.OutputVars {
 			outVars = append(outVars, outVar.Name)
 		}
 		//cleaning the directory
 		err = os.RemoveAll(util.Output_path)
 		if err != nil {
 			log.Println(util.DEVTRON, err)
-			return nil, ciStep, err
+			return nil, step, err
 		}
 		err = os.MkdirAll(util.Output_path, os.ModePerm|os.ModeDir)
 		if err != nil {
 			log.Println(util.DEVTRON, err)
-			return nil, ciStep, err
+			return nil, step, err
 		}
 
 		stepOutputVarsFinal := make(map[string]string)
 		//---------------------------------------------------------------------------------------------------
-		if ciStep.StepType == helper.STEP_TYPE_INLINE {
-			if ciStep.ExecutorType == helper.SHELL {
-				stageOutputVars, err := RunScripts(util.Output_path, fmt.Sprintf("stage-%d", i), ciStep.Script, scriptEnvs, outVars)
+		if step.StepType == helper.STEP_TYPE_INLINE {
+			if step.ExecutorType == helper.SHELL {
+				stageOutputVars, err := RunScripts(util.Output_path, fmt.Sprintf("stage-%d", i), step.Script, scriptEnvs, outVars)
 				if err != nil {
-					return nil, ciStep, err
+					return nil, step, err
 				}
 				stepOutputVarsFinal = stageOutputVars
-				if len(ciStep.ArtifactPaths) > 0 {
-					for _, path := range ciStep.ArtifactPaths {
-						err = copy.Copy(path, filepath.Join(util.TmpArtifactLocation, ciStep.Name, path))
+				if len(step.ArtifactPaths) > 0 {
+					for _, path := range step.ArtifactPaths {
+						err = copy.Copy(path, filepath.Join(util.TmpArtifactLocation, step.Name, path))
 						if err != nil {
 							if _, ok := err.(*os.PathError); ok {
 								log.Println(util.DEVTRON, "dir not exists", path)
 								continue
 							} else {
-								return nil, ciStep, err
+								return nil, step, err
 							}
 						}
 					}
 				}
-			} else if ciStep.ExecutorType == helper.CONTAINER_IMAGE {
+			} else if step.ExecutorType == helper.CONTAINER_IMAGE {
 				var outputDirMount []*helper.MountPath
 				stepArtifact := filepath.Join(util.Output_path, "opt")
 
-				for _, artifact := range ciStep.ArtifactPaths {
+				for _, artifact := range step.ArtifactPaths {
 					hostPath := filepath.Join(stepArtifact, artifact)
 					err = os.MkdirAll(hostPath, os.ModePerm|os.ModeDir)
 					if err != nil {
 						log.Println(util.DEVTRON, err)
-						return nil, ciStep, err
+						return nil, step, err
 					}
 					path := &helper.MountPath{DstPath: artifact, SrcPath: filepath.Join(stepArtifact, artifact)}
 					outputDirMount = append(outputDirMount, path)
 				}
 				executionConf := &executionConf{
-					Script:            ciStep.Script,
+					Script:            step.Script,
 					EnvInputVars:      scriptEnvs,
-					ExposedPorts:      ciStep.ExposedPorts,
+					ExposedPorts:      step.ExposedPorts,
 					OutputVars:        outVars,
-					DockerImage:       ciStep.DockerImage,
-					command:           ciStep.Command,
-					args:              ciStep.Args,
-					CustomScriptMount: ciStep.CustomScriptMount,
-					SourceCodeMount:   ciStep.SourceCodeMount,
-					ExtraVolumeMounts: ciStep.ExtraVolumeMounts,
+					DockerImage:       step.DockerImage,
+					command:           step.Command,
+					args:              step.Args,
+					CustomScriptMount: step.CustomScriptMount,
+					SourceCodeMount:   step.SourceCodeMount,
+					ExtraVolumeMounts: step.ExtraVolumeMounts,
 					scriptFileName:    fmt.Sprintf("stage-%d", i),
 					workDirectory:     util.Output_path,
 					OutputDirMount:    outputDirMount,
@@ -154,23 +154,23 @@ func RunCiSteps(stepType StepType, steps []*helper.StepObject, refStageMap map[i
 				}
 				stageOutputVars, err := RunScriptsInDocker(executionConf)
 				if err != nil {
-					return nil, ciStep, err
+					return nil, step, err
 				}
 				stepOutputVarsFinal = stageOutputVars
 				if _, err := os.Stat(stepArtifact); os.IsNotExist(err) {
 					// Ignore if no file/folder
 					log.Println(util.DEVTRON, "artifact not found ", err)
 				} else {
-					err = copy.Copy(stepArtifact, filepath.Join(util.TmpArtifactLocation, ciStep.Name))
+					err = copy.Copy(stepArtifact, filepath.Join(util.TmpArtifactLocation, step.Name))
 					if err != nil {
-						return nil, ciStep, err
+						return nil, step, err
 					}
 				}
 			}
-		} else if ciStep.StepType == helper.STEP_TYPE_REF_PLUGIN {
-			steps := refStageMap[ciStep.RefPluginId]
+		} else if step.StepType == helper.STEP_TYPE_REF_PLUGIN {
+			steps := refStageMap[step.RefPluginId]
 			stepIndexVarNameValueMap := make(map[int]map[string]string)
-			for _, inVar := range ciStep.InputVars {
+			for _, inVar := range step.InputVars {
 				if varMap, ok := stepIndexVarNameValueMap[inVar.VariableStepIndexInPlugin]; ok {
 					varMap[inVar.Name] = inVar.Value
 					stepIndexVarNameValueMap[inVar.VariableStepIndexInPlugin] = varMap
@@ -188,12 +188,12 @@ func RunCiSteps(stepType StepType, steps []*helper.StepObject, refStageMap map[i
 					}
 				}
 			}
-			opt, _, err := RunCiSteps(STEP_TYPE_REF_PLUGIN, steps, refStageMap, globalEnvironmentVariables, nil)
+			opt, _, err := RunCiCdSteps(STEP_TYPE_REF_PLUGIN, steps, refStageMap, globalEnvironmentVariables, nil)
 			if err != nil {
 				fmt.Println(err)
-				return nil, ciStep, err
+				return nil, step, err
 			}
-			for _, outputVar := range ciStep.OutputVars {
+			for _, outputVar := range step.OutputVars {
 				if varObj, ok := opt[outputVar.VariableStepIndexInPlugin]; ok {
 					if v, ok1 := varObj[outputVar.Name]; ok1 {
 						stepOutputVarsFinal[v.Name] = v.Value
@@ -206,28 +206,28 @@ func RunCiSteps(stepType StepType, steps []*helper.StepObject, refStageMap map[i
 			// artifact path
 			//
 		} else {
-			return nil, ciStep, fmt.Errorf("step Type :%s not supported", ciStep.StepType)
+			return nil, step, fmt.Errorf("step Type :%s not supported", step.StepType)
 		}
 		//---------------------------------------------------------------------------------------------------
-		finalOutVars, err := populateOutVars(stepOutputVarsFinal, ciStep.OutputVars)
+		finalOutVars, err := populateOutVars(stepOutputVarsFinal, step.OutputVars)
 		if err != nil {
-			return nil, ciStep, err
+			return nil, step, err
 		}
-		ciStep.OutputVars = finalOutVars
-		if len(ciStep.SuccessFailureConditions) > 0 {
-			success, err := helper.StageIsSuccess(ciStep.SuccessFailureConditions, finalOutVars)
+		step.OutputVars = finalOutVars
+		if len(step.SuccessFailureConditions) > 0 {
+			success, err := helper.StageIsSuccess(step.SuccessFailureConditions, finalOutVars)
 			if err != nil {
-				return nil, ciStep, err
+				return nil, step, err
 			}
 			if !success {
-				return nil, ciStep, fmt.Errorf("stage not successful because of condition failure")
+				return nil, step, fmt.Errorf("stage not successful because of condition failure")
 			}
 		}
 		finalOutVarMap := make(map[string]*helper.VariableObject)
-		for _, out := range ciStep.OutputVars {
+		for _, out := range step.OutputVars {
 			finalOutVarMap[out.Name] = out
 		}
-		stageVariable[ciStep.Index] = finalOutVarMap
+		stageVariable[step.Index] = finalOutVarMap
 	}
 	return stageVariable, nil, nil
 }
@@ -316,83 +316,83 @@ func deduceVariables(desiredVars []*helper.VariableObject, globalVars map[string
 
 }
 
-func RunPreDockerBuildTasks(ciRequest *helper.CiRequest, scriptEnvs map[string]string, taskYaml *helper.TaskYaml) error {
-	//before task
-	beforeTaskMap := make(map[string]*helper.Task)
-	for i, task := range ciRequest.BeforeDockerBuild {
-		task.RunStatus = true
-		beforeTaskMap[task.Name] = task
-		log.Println(util.DEVTRON, "pre", task)
-		//log running cmd
-		util.LogStage(task.Name)
-		_, err := RunScripts(util.Output_path, fmt.Sprintf("before-%d", i), task.Script, scriptEnvs, nil)
-		if err != nil {
-			return err
-		}
-	}
+//func RunPreDockerBuildTasks(ciRequest *helper.CiRequest, scriptEnvs map[string]string, taskYaml *helper.TaskYaml) error {
+//	//before task
+//	beforeTaskMap := make(map[string]*helper.Task)
+//	for i, task := range ciRequest.BeforeDockerBuild {
+//		task.RunStatus = true
+//		beforeTaskMap[task.Name] = task
+//		log.Println(util.DEVTRON, "pre", task)
+//		//log running cmd
+//		util.LogStage(task.Name)
+//		_, err := RunScripts(util.Output_path, fmt.Sprintf("before-%d", i), task.Script, scriptEnvs, nil)
+//		if err != nil {
+//			return err
+//		}
+//	}
+//
+//	beforeYamlTasks, err := helper.GetBeforeDockerBuildTasks(ciRequest, taskYaml)
+//	if err != nil {
+//		log.Println(err)
+//		return err
+//	}
+//
+//	// run before yaml tasks
+//	for i, task := range beforeYamlTasks {
+//		if _, ok := beforeTaskMap[task.Name]; ok {
+//			log.Println("duplicate task found in yaml, ran earlier so ignoring")
+//			continue
+//		}
+//		beforeTaskMap[task.Name] = task
+//		task.RunStatus = true
+//		log.Println(util.DEVTRON, "pre - yaml", task)
+//		//log running cmd
+//		util.LogStage(task.Name)
+//		_, err = RunScripts(util.Output_path, fmt.Sprintf("before-yaml-%d", i), task.Script, scriptEnvs, nil)
+//		if err != nil {
+//			return err
+//		}
+//	}
+//	return nil
+//}
 
-	beforeYamlTasks, err := helper.GetBeforeDockerBuildTasks(ciRequest, taskYaml)
-	if err != nil {
-		log.Println(err)
-		return err
-	}
-
-	// run before yaml tasks
-	for i, task := range beforeYamlTasks {
-		if _, ok := beforeTaskMap[task.Name]; ok {
-			log.Println("duplicate task found in yaml, ran earlier so ignoring")
-			continue
-		}
-		beforeTaskMap[task.Name] = task
-		task.RunStatus = true
-		log.Println(util.DEVTRON, "pre - yaml", task)
-		//log running cmd
-		util.LogStage(task.Name)
-		_, err = RunScripts(util.Output_path, fmt.Sprintf("before-yaml-%d", i), task.Script, scriptEnvs, nil)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func RunPostDockerBuildTasks(ciRequest *helper.CiRequest, scriptEnvs map[string]string, taskYaml *helper.TaskYaml) error {
-	log.Println(util.DEVTRON, " docker-build-post-processing")
-	afterTaskMap := make(map[string]*helper.Task)
-	for i, task := range ciRequest.AfterDockerBuild {
-		task.RunStatus = true
-		afterTaskMap[task.Name] = task
-		log.Println(util.DEVTRON, "post", task)
-		util.LogStage(task.Name)
-		_, err := RunScripts(util.Output_path, fmt.Sprintf("after-%d", i), task.Script, scriptEnvs, nil)
-		if err != nil {
-			return err
-		}
-	}
-
-	afterYamlTasks, err := helper.GetAfterDockerBuildTasks(ciRequest, taskYaml)
-	if err != nil {
-		log.Println(err)
-		return err
-	}
-
-	for i, task := range afterYamlTasks {
-		if _, ok := afterTaskMap[task.Name]; ok {
-			log.Println("duplicate task found in yaml, already run so ignoring")
-			continue
-		}
-		afterTaskMap[task.Name] = task
-		task.RunStatus = true
-		log.Println(util.DEVTRON, "post - yaml", task)
-		//log running cmd
-		util.LogStage(task.Name)
-		_, err = RunScripts(util.Output_path, fmt.Sprintf("after-yaml-%d", i), task.Script, scriptEnvs, nil)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
+//func RunPostDockerBuildTasks(ciRequest *helper.CiRequest, scriptEnvs map[string]string, taskYaml *helper.TaskYaml) error {
+//	log.Println(util.DEVTRON, " docker-build-post-processing")
+//	afterTaskMap := make(map[string]*helper.Task)
+//	for i, task := range ciRequest.AfterDockerBuild {
+//		task.RunStatus = true
+//		afterTaskMap[task.Name] = task
+//		log.Println(util.DEVTRON, "post", task)
+//		util.LogStage(task.Name)
+//		_, err := RunScripts(util.Output_path, fmt.Sprintf("after-%d", i), task.Script, scriptEnvs, nil)
+//		if err != nil {
+//			return err
+//		}
+//	}
+//
+//	afterYamlTasks, err := helper.GetAfterDockerBuildTasks(ciRequest, taskYaml)
+//	if err != nil {
+//		log.Println(err)
+//		return err
+//	}
+//
+//	for i, task := range afterYamlTasks {
+//		if _, ok := afterTaskMap[task.Name]; ok {
+//			log.Println("duplicate task found in yaml, already run so ignoring")
+//			continue
+//		}
+//		afterTaskMap[task.Name] = task
+//		task.RunStatus = true
+//		log.Println(util.DEVTRON, "post - yaml", task)
+//		//log running cmd
+//		util.LogStage(task.Name)
+//		_, err = RunScripts(util.Output_path, fmt.Sprintf("after-yaml-%d", i), task.Script, scriptEnvs, nil)
+//		if err != nil {
+//			return err
+//		}
+//	}
+//	return nil
+//}
 
 func RunCdStageTasks(tasks []*helper.Task, scriptEnvs map[string]string) error {
 	log.Println(util.DEVTRON, " cd-stage-processing")
