+++ helper/EventHelper.go
@@ -178,6 +178,8 @@ type CdRequest struct {
 	CiRunnerDockerMtuValue     int                               `json:"ciRunnerDockerMtuValue"`
 	DeploymentReleaseCounter   int                               `json:"deploymentReleaseCounter,omitempty"`
 	IsDryRun                   bool                              `json:"isDryRun"`
+	PrePostDeploySteps         []*StepObject                     `json:"prePostDeploySteps"`
+	RefPlugins                 []*RefPluginObject                `json:"refPlugins"`
 }
 
 type CiCdTriggerEvent struct {
